Moved dropzone to UI
